Refactor invoice logic and add ISR simulation service
Moved InvoiceUser classes to a dedicated package and extracted related logic into InvoiceUserService. Refactored InvoiceService to delegate details and user handling to new services, simplifying controller logic. Added TaxIsrCalculationService for ISR simulation and updated IsrSimulationRestController to use it. Improved LLMService tax rate handling with TaxRateEnum. Cleaned up and streamlined REST controllers for invoices and XML processing.
