+++ src/main/java/com/project/demo/logic/entity/invoiceUser/InvoiceUserRepository.java
@@ -1,9 +1,7 @@
-package com.project.demo.logic.entity.invoice;
+package com.project.demo.logic.entity.invoiceUser;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InvoiceUserRepository extends JpaRepository<InvoiceUser, Long> {
-
-
     InvoiceUser findByIdentification(String identification);
 }
