+++ src/main/java/com/project/demo/logic/entity/xml/XmlService.java
@@ -2,7 +2,7 @@ package com.project.demo.logic.entity.xml;
 
 import com.project.demo.logic.entity.detailsInvoice.DetailsInvoice;
 import com.project.demo.logic.entity.invoice.Invoice;
-import com.project.demo.logic.entity.invoice.InvoiceUser;
+import com.project.demo.logic.entity.invoiceUser.InvoiceUser;
 import org.springframework.stereotype.Service;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
